requirements.txt
# fogo_companion/requirements.txt

flask
python-dotenv
pypdf
python-docx
requests
werkzeug
gunicorn